Move FileExtensionSuffixIdentifier to package-agnostic-utilities
