+++ packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/serializableErrorVoictent.ts
@@ -1,7 +1,7 @@
 import { Gepp } from '../../core/engine-shell/voictent/gepp';
 import { Voictent2 } from '../../core/engine/voictent2';
 import { Voque } from '../../core/engine/voque';
-import { FileExtensionSuffixIdentifier } from '../../adapted-programs/programmable-units/file/fileExtensionSuffixIdentifier';
+import { FileExtensionSuffixIdentifier } from '../../package-agnostic-utilities/file/fileExtensionSuffixIdentifier';
 import { ProgramFileCache } from '../../layer-agnostic-utilities/program/programFileCache';
 
 export type SerializableErrorVoque<TGepp extends Gepp> = Voque<
