+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/file/fileExtensionSuffixIdentifier.ts
@@ -1,4 +1,4 @@
-import { swapEntries } from '../../../package-agnostic-utilities/object/swapEntries';
+import { swapEntries } from '../object/swapEntries';
 
 export enum FileExtensionSuffixIdentifier {
   Bash = 'Bash',
